issue #336: lseek / pread64 / pwrite64 / copy_file_range の off_t を long 化 (2GB 超ファイル)
FileAccess.FileSeek が offset / 内部位置 (o/size/curptr) / 戻り値を int で持ち、
RandomAccessFile の getFilePointer()/length() を (int) に丸めていたため、off_t
(64-bit) が 2GB (2^31) 境界で切り詰められていた。lseek だけでなく FileSeek を
経由する pread64 / pwrite64 / copy_file_range も同じ欠陥を共有 (各所 (int)offset)。

修正:
- FileSeek を long FileSeek(int fd, long offset, int whence) に。内部 o/size/curptr
  を long、getFilePointer()/length() の (int) cast を除去、seek(long) をそのまま使用。
  dir 分岐の set_ptr(int) には (int)o (dir cookie は小なので可)。
- sys_lseek (Syscall) / amd64_pread64 / amd64_pwrite64 / amd64_copy_file_range の
  (int)offset / (int)in_off / (int)out_off cast を除去、saved/in_saved/out_saved を long に。

回帰: 新規 sys_lseek_large64 (4GiB+5 で lseek SET/END + write/read roundtrip +
pwrite64/pread64 + hole 確認、sparse file) PASS。sys_lseek64 / sys_pwrite64 /
sys_copy_file_range64 / hello_dyn64 (ld.so の libc file-backed mmap = 内部 FileSeek) /
fileio64 / mmap_dyn64 PASS。

残 (#336 内): mmap の file offset 経路 (alloc_and_map の int offset / AllocInfo.map_offset)
は MemoryBackend interface 波及のため別 commit/PR とする。

+++ src/main/java/emulin/FileAccess.java
@@ -398,11 +398,13 @@ public class FileAccess
   }
 
   // ファイルシーク
-  int FileSeek( int fd, int offset, int whence ) {
+  // issue #336: off_t は 64-bit。offset / 内部位置 / 戻り値を long で扱う
+  //   (旧 int 実装は 2GB(2^31) 境界で seek 位置・size・返り値を切り詰めていた)。
+  long FileSeek( int fd, long offset, int whence ) {
     boolean ret = true;
-    int o = 0;
-    int size = 0;
-    int curptr = 0;
+    long o = 0;
+    long size = 0;
+    long curptr = 0;
     // issue #334 (apt install 中に露呈): 無効 fd は EBADF(-9) を返す。旧実装は fd が
     //   範囲外なら elementAt で AIOOBE、finfo==null なら print 後に下の finfo.f 参照で
     //   NPE となり、apt の worker thread が落ちて親が wait4 で永久 block していた
@@ -432,9 +434,9 @@ public class FileAccess
       }
     }
     if( null != finfo.f ) { // ディレクトリ以外なら実際のシークを実行する
-      try { curptr = (int)finfo.f.getFilePointer( ); }
+      try { curptr = finfo.f.getFilePointer( ); }   // issue #336: long (旧 (int) 切り詰め)
       catch ( IOException m ) { ret = false; }
-      try { size = (int)finfo.f.length( ); }
+      try { size = finfo.f.length( ); }              // issue #336: long (旧 (int) 切り詰め)
       catch ( IOException m ) { ret = false; }
       if( ret ) {
 	if( whence == SEEK_SET ) {
@@ -461,7 +463,7 @@ public class FileAccess
       if( whence == SEEK_END ) {
 	o = size + offset;
       }
-      set_ptr( fd, o );
+      set_ptr( fd, (int)o );   // issue #336: dir cookie は小さいので (int) で可
     }
     if( ret ) {
       return( o );
